api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["jp calvo <calvojp92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
sqlmodel = "^0.0.4"
psycopg2 = "^2.9.2"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
email-validator = "^1.1.3"
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"
python-jose = "^3.3.0"
redis = "^4.1.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.93.14"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.12.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
